fix(model): bedrooms and bathrooms are not integers — 7.5 baths broke the whole page payload
Casa Luar has 7.5 baths. `Listing.baths` was `Option<i64>`, so serde rejected the ENTIRE page payload on that one field:

    invalid type: floating point `7.5`, expected i64

`page_loaded_json` turned that into an effect failure, the model kept `page: None`, and `site_home` — which renders
nothing without a page — produced no hero, no portfolio, no services, no culture and no about. A single half-bath
removed the whole page, and the failure looked like a rendering problem rather than a data one.

Found by piping the real payload through the view and reading what came out — `view::render` printed 0 bytes and named
the field. That is the check I should have run before the first reload instead of after the fourth: the view is pure, so
it can be rendered from the live payload outside a browser, and "0 bytes" is unambiguous where a white page is not.

Both counts are `f64` now, and a whole number prints without a decimal point: "8 Bed", "7.5 Bath", "1 Acre".

Verified against the running server: the page renders 20,177 bytes with every section — top, properties, buyers,
sellers, culture, about — and the real content in it: `Casa Luar` six times, `$2,500,000`, the hero photograph, and
the half-bath.

+++ rust/ui/src/model.rs
@@ -339,8 +339,11 @@ pub struct Listing {
     pub kind: Option<String>,
     pub image_path: Option<String>,
     pub image_alt: Option<String>,
-    pub beds: Option<i64>,
-    pub baths: Option<i64>,
+    /// Bedrooms and bathrooms are NOT integers. Half-baths are ordinary, so `7.5` is a real value in this data — and
+    /// declaring these as `i64` made the whole page payload fail to deserialize the first time a listing had one, which
+    /// presents as a page with no hero and no sections rather than as a bad number.
+    pub beds: Option<f64>,
+    pub baths: Option<f64>,
     pub area: Option<String>,
     /// Whether the estate is in the featured set, which is what draws the badge on its card.
     pub featured: bool,
